# internal/models/tasks.go
package models

import (
	"time"

	"go.mongodb.org/mongo-driver/bson/primitive"
)
type Task struct {
	ID primitive.ObjectID `bson:"_id,omitempty" json:"id"`
	Title string `bson:"title" json:"title"`
	Completed bool               `bson:"completed" json:"completed"`
	CreatedAt time.Time          `bson:"created_at" json:"created_at"`
}